feat: model-agnostic binary engine with universal quant support (#436)
* fix: resolve 18 remaining open issues (HIGH: #420-#422, #434; MEDIUM: #419, #423-#430, #435; LOW: #389, #399, #431-#433)

HIGH severity fixes:
- #434: Set npu_xrt auto_selectable=false until kernel correctness verified
- #422: Add input validation in run_albert_opt.c to prevent command injection
- #421: Fix fprintf missing argument (UB) in npu_engine_cb.cpp
- #420: Fix EPS=1e-6f truncated to 0 by int declaration in npu_engine_hybrid.cpp

MEDIUM severity fixes:
- #435: Default bind to 127.0.0.1 instead of 0.0.0.0, add --bind CLI flag
- #430: Fix malformed snprintf in test_parallel_real.hip
- #428: Fix integer overflow in npu_ternaryd.cpp size calculations
- #427: Fix pipe fd leaks on partial failure in backend_npu.cpp x2
- #426: Fix TOCTOU race in BackendManager health_check/failover
- #425: Fix sizeof mismatch skip in gguf_loader.cpp
- #424: Add fread return checks and token_id OOB bounds
- #423: Add division-by-zero guards in model loaders
- #419: Fix GGUF parser skipping value data for long keys and tokenizer.ggml.model
- #394: Add runtime validation of model dims in backend_hip.cpp

LOW severity fixes:
- #433: Remove local HIP_OK_V definition, use hip_check.h
- #432: Fix compilation warnings (unused vars, missing initializers, shadowed names)
- #431: Shell script hardening (consistent set -e, remove double shebangs)
- #399: KV cache PartialsCache proper deallocation
- #389: Replace release-unsafe assert() with runtime checks in buffer.hpp

Closed as not-actionable:
- #20: Windows NPU testing (community)
- #21: macOS ZINC testing (community)

* feat: model-agnostic binary engine with universal quant support

- HIP backend: refuse non-Zaya models, let token router fallback (no crashes)
- Token router: reorder backends: CPU before NPU FLM for correct fallback
- K-quant dequant: Q2_K, Q3_K, Q4_K, Q5_K, Q6_K, Q8_K (all superblock formats)
- Legacy quants: Q4_1, Q5_0, Q5_1 dequant added
- Integer types: I8, I16, I32 raw tensor support
- GGUF reader: reject unsupported quants instead of zero-fill
- HIP backend: fix missing d_prev_rs/d_expert_* allocations in load_gguf_model
- block_info: fix K-quant type offset mapping (was off-by-2)
- CORS: opt-in via --cors-origin flag (was hardcoded *)
- Shell hygiene: set -euo pipefail on 15 scripts, $HOME paths on 8 scripts
- HIP_CHECK macro: add HIP_OK_R to shared hip_check.h
- zaya_engine: fix [redacted] placeholders -> eng.vocab bounds checks
- backend_hip: extract dims to zaya1_dims.h with #ifndef guards
- tests: 13/13 pass, both targets build clean

Closes #290, #376-#379, #382, #384-#389, #394, #399, #418-#435

---------

+++ engine/npu/src/npu_ternaryd.cpp
@@ -204,11 +204,11 @@ struct TernaryCtx {
         sc_uuid = single_uuid;  sc_instr = single_instr;
         mc_uuid = multi_uuid;   mc_instr = multi_instr;
 
-        weights_bytes = (size_t)M_total * K_packed_total;
-        scales_bytes = (size_t)M_total * n_blocks * 2;
+        weights_bytes = (size_t)M_total * (size_t)K_packed_total;
+        scales_bytes = (size_t)M_total * (size_t)n_blocks * 2;
         // Keep a HOST copy of weights/scales for reliable access
         weights_host_ptr = new uint8_t[weights_bytes];
-        scales_host_ptr = new uint16_t[(size_t)M_total * n_blocks];
+        scales_host_ptr = new uint16_t[(size_t)M_total * (size_t)n_blocks];
         memcpy((void*)weights_host_ptr, packed_weights, weights_bytes);
         memcpy((void*)scales_host_ptr, weight_scales, scales_bytes);
         // Also create device BO for NPU access
